Changed a u8 to a u16 that was causing OBJ_GFX issues in Multi Battle

+++ src/battle_tower.c
@@ -1163,7 +1163,7 @@ void SetBattleFacilityTrainerGfxId(u16 trainerId, u8 tempVarId)
 {
     u32 i;
     u8 facilityClass;
-    u8 trainerObjectGfxId;
+    u16 trainerObjectGfxId;
 
     SetFacilityPtrsGetLevel();
     if (trainerId == TRAINER_EREADER)
